cleanup and finish major refactoring of handlers
